kodilla-selenium-login/src/main/java/com/kodilla/selenium/pom/KodillaLoginPom.java
package com.kodilla.selenium.pom;

import org.openqa.selenium.WebDriver;
import org.openqa.selenium.WebElement;
import org.openqa.selenium.support.FindBy;
import org.openqa.selenium.support.PageFactory;

public class KodillaLoginPom extends AbstractPom {

    @FindBy(css = "input[type='email']")
    private WebElement emailField;

    @FindBy(css = "input[type='password']")
    private WebElement passwordField;

    @FindBy(xpath = "//button[contains(text(), 'Log in')]")
    private WebElement loginButton;

    public KodillaLoginPom(WebDriver driver) {
        super(driver);
        PageFactory.initElements(driver, this);
    }

    public boolean login(String email, String password) {
        emailField.sendKeys(email);
        passwordField.sendKeys(password);
        loginButton.click();
        String message = driver.switchTo().alert().getText();
        driver.switchTo().alert().dismiss();
        return message.equals("Jesteś teraz zalogowany!");
    }
}
